Reduce the number of specialised processes; less bulk, more clarity
