+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/transceiver/processes/MulticastRoutesControlProcess.java
@@ -16,25 +16,50 @@
  */
 package uk.ac.manchester.spinnaker.transceiver.processes;
 
+import static java.lang.Byte.toUnsignedInt;
+import static java.lang.Integer.toUnsignedLong;
 import static java.nio.ByteBuffer.allocate;
 import static java.nio.ByteOrder.LITTLE_ENDIAN;
+import static uk.ac.manchester.spinnaker.machine.MachineDefaults.ROUTER_AVAILABLE_ENTRIES;
+import static uk.ac.manchester.spinnaker.messages.Constants.UDP_MESSAGE_MAX_SIZE;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.TreeMap;
 
 import uk.ac.manchester.spinnaker.connections.SCPConnection;
 import uk.ac.manchester.spinnaker.connections.selectors.ConnectionSelector;
 import uk.ac.manchester.spinnaker.machine.HasChipLocation;
 import uk.ac.manchester.spinnaker.machine.MulticastRoutingEntry;
 import uk.ac.manchester.spinnaker.messages.model.AppID;
+import uk.ac.manchester.spinnaker.messages.scp.ReadMemory;
 import uk.ac.manchester.spinnaker.messages.scp.RouterAlloc;
 import uk.ac.manchester.spinnaker.messages.scp.RouterInit;
 import uk.ac.manchester.spinnaker.transceiver.RetryTracker;
 
-/** A process for loading the multicast routing table on a SpiNNaker chip. */
-public class LoadMulticastRoutesProcess
+/**
+ * A process for reading and writing the multicast routing table of a SpiNNaker
+ * chip.
+ */
+public class MulticastRoutesControlProcess
 		extends MultiConnectionProcess<SCPConnection> {
+	private static final long INVALID_ROUTE_MARKER = 0xFF000000L;
+	/** Each routing table entry is 16 bytes long. */
+	private static final int BYTES_PER_ENTRY = 16;
+	/** 16 entries fit in a 256-byte read. */
+	private static final int ENTRIES_PER_READ =
+			UDP_MESSAGE_MAX_SIZE / BYTES_PER_ENTRY;
+	/** 64 reads of 16 entries are required for 1024 entries. */
+	private static final int NUM_READS =
+			ROUTER_AVAILABLE_ENTRIES / ENTRIES_PER_READ;
+	private static final int UNIT_SIZE = 16;
+	private static final int END = 0xFFFFFFFF;
+	private static final int ROUTING_TABLE_ADDRESS = 0x67800000;
+
 	private RetryTracker retryTracker;
 
 	/**
@@ -45,16 +70,12 @@ public class LoadMulticastRoutesProcess
 	 *            operation. May be {@code null} if no suck tracking is
 	 *            required.
 	 */
-	public LoadMulticastRoutesProcess(
+	public MulticastRoutesControlProcess(
 			ConnectionSelector<SCPConnection> connectionSelector,
 			RetryTracker retryTracker) {
 		super(connectionSelector, retryTracker);
 		this.retryTracker = retryTracker;
 	}
-
-	private static final int UNIT_SIZE = 16;
-	private static final int END = 0xFFFFFFFF;
-	private static final int ROUTING_TABLE_ADDRESS = 0x67800000;
 
 	private static void writeEntry(ByteBuffer buffer, short index,
 			MulticastRoutingEntry route) {
@@ -86,7 +107,7 @@ public class LoadMulticastRoutesProcess
 	 * @throws ProcessException
 	 *             If SpiNNaker rejects a message.
 	 */
-	public void loadRoutes(HasChipLocation chip,
+	public void setRoutes(HasChipLocation chip,
 			Collection<MulticastRoutingEntry> routes, AppID appID)
 			throws IOException, ProcessException {
 		/*
@@ -119,5 +140,57 @@ public class LoadMulticastRoutesProcess
 		// Load the entries
 		synchronousCall(new RouterInit(chip, routes.size(),
 				ROUTING_TABLE_ADDRESS, baseAddress, appID));
+	}
+
+	/**
+	 * Get the multicast routes from a chip's router.
+	 *
+	 * @param chip
+	 *            The chip to read from.
+	 * @param baseAddress
+	 *            Where the routing table is.
+	 * @param appID
+	 *            What application is associated with the routes we are
+	 *            interested in. Use {@code null} to read all routes.
+	 * @return The list of routes.
+	 * @throws IOException
+	 *             If anything goes wrong with networking.
+	 * @throws ProcessException
+	 *             If SpiNNaker rejects a message.
+	 */
+	public List<MulticastRoutingEntry> getRoutes(HasChipLocation chip,
+			int baseAddress, AppID appID)
+			throws IOException, ProcessException {
+		Map<Integer, MulticastRoutingEntry> routes = new TreeMap<>();
+		for (int i = 0; i < NUM_READS; i++) {
+			int offset = i * ENTRIES_PER_READ;
+			sendRequest(
+					new ReadMemory(chip, baseAddress + offset * BYTES_PER_ENTRY,
+							UDP_MESSAGE_MAX_SIZE),
+					response -> addRoutes(response.data, offset, routes,
+							appID));
+		}
+		finish();
+		checkForError();
+		return new ArrayList<>(routes.values());
+	}
+
+	private void addRoutes(ByteBuffer data, int offset,
+			Map<Integer, MulticastRoutingEntry> routes, AppID appID) {
+		for (int r = 0; r < ENTRIES_PER_READ; r++) {
+			data.get(); // Ignore
+			data.get(); // Ignore
+			AppID appid = new AppID(toUnsignedInt(data.get()));
+			data.get(); // Ignore
+			int route = data.getInt();
+			int key = data.getInt();
+			int mask = data.getInt();
+
+			if (toUnsignedLong(route) < INVALID_ROUTE_MARKER
+					&& (appID == null || appID.equals(appid))) {
+				routes.put(r + offset,
+						new MulticastRoutingEntry(key, mask, route, false));
+			}
+		}
 	}
 }
