--- SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/transceiver/processes/SaveApplicationRouterTableProcess.java
@@ -1,79 +1,0 @@
-/*
- * Copyright (c) 2019 The University of Manchester
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package uk.ac.manchester.spinnaker.transceiver.processes;
-
-import java.io.IOException;
-
-import uk.ac.manchester.spinnaker.connections.SCPConnection;
-import uk.ac.manchester.spinnaker.connections.selectors.ConnectionSelector;
-import uk.ac.manchester.spinnaker.machine.CoreLocation;
-import uk.ac.manchester.spinnaker.machine.CoreSubsets;
-import uk.ac.manchester.spinnaker.messages.scp.RouterTableSaveApplicationRoutes;
-import uk.ac.manchester.spinnaker.transceiver.RetryTracker;
-
-/** Save the application multicast router tables for a set of routers. */
-public class SaveApplicationRouterTableProcess
-		extends MultiConnectionProcess<SCPConnection> {
-	/**
-	 * @param connectionSelector
-	 *            How to select how to communicate.
-	 * @param retryTracker
-	 *            Object used to track how many retries were used in an
-	 *            operation. May be {@code null} if no suck tracking is
-	 *            required.
-	 */
-	public SaveApplicationRouterTableProcess(
-			ConnectionSelector<SCPConnection> connectionSelector,
-			RetryTracker retryTracker) {
-		super(connectionSelector, retryTracker);
-	}
-
-	/**
-	 * Save the application's multicast router table for a particular router.
-	 *
-	 * @param monitorCore
-	 *            the extra monitor core on the chip with the router.
-	 * @throws IOException
-	 *             If anything goes wrong with networking.
-	 * @throws ProcessException
-	 *             If SpiNNaker rejects the message.
-	 */
-	public void saveApplicationRouterTable(CoreLocation monitorCore)
-			throws IOException, ProcessException {
-		synchronousCall(new RouterTableSaveApplicationRoutes(monitorCore));
-	}
-
-	/**
-	 * Save the application's multicast router tables for a collection of
-	 * routers.
-	 *
-	 * @param monitorCoreSubsets
-	 *            the cores where the monitors are running.
-	 * @throws IOException
-	 *             If anything goes wrong with networking.
-	 * @throws ProcessException
-	 *             If SpiNNaker rejects the message.
-	 */
-	public void saveApplicationRouterTable(CoreSubsets monitorCoreSubsets)
-			throws IOException, ProcessException {
-		for (CoreLocation core : monitorCoreSubsets) {
-			sendRequest(new RouterTableSaveApplicationRoutes(core));
-		}
-		finish();
-		checkForError();
-	}
-}
